refactor: Move admin nav into a wrapper
